Refactor limit variable names for clarity
Variables in the Limits class pertaining to upper and lower limits have been renamed for increased readability and consistency. The 'limit_' prefix was superfluous and has been removed, streamlining the identification of these related variables in the context of their associated calculations and method calls. This change enhances the codebase maintainability.