Consolidate common headers into a single file
Centralized the inclusion of common header files into a newly created .mq5 file, streamlining the reference process across the project. This change simplifies code maintenance and future updates by reducing redundancy and the potential for errors in header file management. Additionally, the introduction of an .mq5 file enhances project organization, making it easier to identify and manage shared dependencies. The replacement of individual header inclusions with a single reference in the .mqh file underscores our commitment to efficient and clean codebase practices. Moreover, the project's structure is further refined by the addition of a new project file (.mqproj) and the removal of the previous AutomatedTrading.mqproj, indicating a reorganization and possibly an upgrade of the project's configuration setup.